requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sentence-transformers==3.4.1
faiss-cpu==1.8.0.post1
pandas==2.2.2
numpy==1.26.4  # Downgraded to be compatible with faiss-cpu
PyPDF2==3.0.1
Pillow==10.4.0
pdfplumber==0.11.4
nltk==3.9.1
pydantic==2.9.2
python-multipart==0.0.9
nest_asyncio==1.6.0
scikit-learn==1.5.2
joblib==1.4.2
threadpoolctl==3.5.0